♻️ [refactor]: tag directory 위치 변경
